+++ projects/lib/badge/src/badge.component.ts
@@ -6,12 +6,19 @@
  * You may not use this file except in compliance with the License
  * that can be found at http://www.apache.org/licenses/LICENSE-2.0
  */
-import {booleanAttribute, ChangeDetectionStrategy, Component, Input, ViewEncapsulation} from "@angular/core";
+import {
+  booleanAttribute,
+  ChangeDetectionStrategy,
+  Component,
+  HostBinding,
+  Input,
+  ViewEncapsulation
+} from "@angular/core";
 import {CuteBaseControl} from "@cute-widgets/base/abstract";
 import {toTextBgCssClass} from "@cute-widgets/base/core";
 
 /**
- * Small count and labeling component
+ * Small count and labeling component.
  */
 @Component({
   selector: 'cute-badge',
@@ -20,8 +27,7 @@ import {toTextBgCssClass} from "@cute-widgets/base/core";
   styleUrls: ['./badge.component.scss'],
   host: {
     'class': 'cute-badge badge',
-    //'[class]' : 'color ? "text-bg-"+color : ""',
-    '[class]': 'color ? toTextBgCssClass(color) : ""',
+    '[class.cute-badge-positioned]': 'position ?? null',
     '[class.rounded-pill]': 'roundedPill',
   },
   standalone: true,
@@ -34,6 +40,28 @@ export class CuteBadge extends CuteBaseControl {
   @Input({transform: booleanAttribute})
   roundedPill: boolean = false;
 
+  /** Badge's position in the corner of parent link or button. */
+  @Input() position: "top-start"|"top-end"|"bottom-start"|"bottom-end" | undefined;
+
+  @HostBinding('class')
+  protected get classList(): string[] {
+    let classes: string[] = [];
+    if (this.color) {
+      classes.push( toTextBgCssClass(this.color) );
+    }
+    if (this.position) {
+      let xClass="", yClass="";
+      switch (this.position) {
+        case "top-start":  xClass = "start-0"; yClass = "top-0"; break;
+        case "top-end":  xClass = "start-100"; yClass = "top-0"; break;
+        case "bottom-start":  xClass = "start-0"; yClass = "top-100"; break;
+        case "bottom-end":  xClass = "start-100"; yClass = "top-100"; break;
+      }
+      classes.push(xClass, yClass);
+    }
+    return classes;
+  }
+
   constructor() {
     super();
   }
@@ -42,6 +70,4 @@ export class CuteBadge extends CuteBaseControl {
     // none
     return "";
   }
-
-  protected readonly toTextBgCssClass = toTextBgCssClass;
 }