+++ projects/showcase/src/pages/components/badge/badge.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import {CuteVStack} from '@cute-widgets/base/layout';
+import {CuteHStack, CuteVStack} from '@cute-widgets/base/layout';
 import {CuteBadge} from '@cute-widgets/base/badge';
 import {CuteButton} from '@cute-widgets/base/button';
 import {AbstractPage} from '../abstract/abstract-page';
@@ -7,12 +7,13 @@ import {ComponentViewer} from "../../component-viewer/component-viewer";
 
 @Component({
   selector: 'app-badge',
-    imports: [
-        CuteVStack,
-        CuteBadge,
-        CuteButton,
-        ComponentViewer
-    ],
+  imports: [
+    CuteVStack,
+    CuteBadge,
+    CuteButton,
+    ComponentViewer,
+    CuteHStack
+  ],
   templateUrl: './badge.html',
   styleUrl: './badge.scss',
 })